requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.9.0
mysql-connector-python==8.0.28
